Update smoke-test-cpp.sh
Using `source` instead of directly calling the script.

The difference is highlighted in the documentation of  `bash` :
```
source filename [arguments]
Read and execute commands from filename in the current shell environment and return the exit status of the last command executed from filename. [...]
```
while directly executing the script, execute it in a child process, so that all environment variables are lost when the script returns.